requirements.txt

python-dotenv
praw
openai
PyYAML
requests
streamlit

